backend/src/requirements.txt
boto3
ffmpeg-python
requests